requirements.txt
google-adk
google-genai
python-dotenv
requests
pillow
markdown
psycopg2-binary
